refactor: update TrapFocus name in docs folder.
