+++ docs/data/base/components/focus-trap/BasicFocusTrap.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import Box from '@mui/system/Box';
-import TrapFocus from '@mui/base/TrapFocus';
+import FocusTrap from '@mui/base/FocusTrap';
 
-export default function LazyTrapFocus() {
+export default function BasicFocusTrap() {
   const [open, setOpen] = React.useState(false);
 
   return (
@@ -11,7 +11,7 @@ export default function LazyTrapFocus() {
         Open
       </button>
       {open && (
-        <TrapFocus open disableAutoFocus>
+        <FocusTrap open>
           <Box tabIndex={-1} sx={{ mt: 1, p: 1 }}>
             <label>
               First name: <input type="text" />
@@ -21,7 +21,7 @@ export default function LazyTrapFocus() {
               Close
             </button>
           </Box>
-        </TrapFocus>
+        </FocusTrap>
       )}
     </Box>
   );
