+++ docs/data/base/components/focus-trap/DisableEnforceFocus.tsx
@@ -1,11 +1,9 @@
 import * as React from 'react';
 import Box from '@mui/system/Box';
-import Portal from '@mui/base/Portal';
-import TrapFocus from '@mui/base/TrapFocus';
+import FocusTrap from '@mui/base/FocusTrap';
 
-export default function PortalTrapFocus() {
+export default function DisableEnforceFocus() {
   const [open, setOpen] = React.useState(false);
-  const [container, setContainer] = React.useState<HTMLElement | null>(null);
 
   return (
     <Box sx={{ display: 'flex', alignItems: 'center', flexDirection: 'column' }}>
@@ -13,25 +11,18 @@ export default function PortalTrapFocus() {
         Open
       </button>
       {open && (
-        <TrapFocus open>
+        <FocusTrap disableEnforceFocus open>
           <Box tabIndex={-1} sx={{ mt: 1, p: 1 }}>
             <label>
               First name: <input type="text" />
             </label>
             <br />
-            <Portal container={container}>
-              <label>
-                Last name: <input type="text" />
-              </label>
-              <br />
-            </Portal>
             <button type="button" onClick={() => setOpen(false)}>
               Close
             </button>
           </Box>
-        </TrapFocus>
+        </FocusTrap>
       )}
-      <div ref={setContainer} />
     </Box>
   );
 }
