+++ docs/data/base/components/focus-trap/PortalFocusTrap.tsx
@@ -1,9 +1,11 @@
 import * as React from 'react';
 import Box from '@mui/system/Box';
-import TrapFocus from '@mui/base/TrapFocus';
+import Portal from '@mui/base/Portal';
+import FocusTrap from '@mui/base/FocusTrap';
 
-export default function BasicTrapFocus() {
+export default function PortalFocusTrap() {
   const [open, setOpen] = React.useState(false);
+  const [container, setContainer] = React.useState<HTMLElement | null>(null);
 
   return (
     <Box sx={{ display: 'flex', alignItems: 'center', flexDirection: 'column' }}>
@@ -11,18 +13,25 @@ export default function BasicTrapFocus() {
         Open
       </button>
       {open && (
-        <TrapFocus open>
+        <FocusTrap open>
           <Box tabIndex={-1} sx={{ mt: 1, p: 1 }}>
             <label>
               First name: <input type="text" />
             </label>
             <br />
+            <Portal container={container}>
+              <label>
+                Last name: <input type="text" />
+              </label>
+              <br />
+            </Portal>
             <button type="button" onClick={() => setOpen(false)}>
               Close
             </button>
           </Box>
-        </TrapFocus>
+        </FocusTrap>
       )}
+      <div ref={setContainer} />
     </Box>
   );
 }
